make_bot/requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
Cython==3.0.11
Django==5.1
filters==1.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
nltk==3.9.1
numpy==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
regex==2024.7.24
requests==2.32.3
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
textblob==0.18.0.post0
tqdm==4.66.5
urllib3==2.2.2
wheel==0.44.0
